server/requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_jwt_extended
psycopg2-binary
python-dotenv